+++ packages/components/src/components/sort-handle/sort-handle.tsx
@@ -4,7 +4,7 @@ import { LitElement, property, createEvent, h, method, JsxNode } from "@arcgis/l
 import { Scale } from "../interfaces";
 import {
   FlipPlacement,
-  MenuPlacement,
+  LogicalPlacement,
   OverlayPositioning,
   defaultMenuPlacement,
 } from "../../utils/floating-ui";
@@ -47,15 +47,68 @@ export class SortHandle extends LitElement {
   }
 
   get hasValidSetInfo(): boolean {
-    return this.hasSetInfo ? this.setPosition > 0 && this.setSize > 1 : true;
+    return this.hasSetInfo
+      ? this.setPosition > 0 && this.setPosition <= this.setSize && this.setSize > 0
+      : true;
   }
 
   get hasReorderItems(): boolean {
     return !this.sortDisabled && this.hasValidSetInfo;
   }
 
+  get hasMoveToItems(): boolean {
+    return this.moveToItems.length > 0;
+  }
+
+  get hasAddToItems(): boolean {
+    return this.addToItems.length > 0;
+  }
+
+  get reorderGroupTitle(): string {
+    return this.hasMoveToItems || this.hasAddToItems ? this.messages.reorder : "";
+  }
+
   get hasNoItems(): boolean {
-    return !this.hasReorderItems && this.moveToItems.length < 1 && this.addToItems.length < 1;
+    return !this.hasReorderItems && !this.hasMoveToItems && !this.hasAddToItems;
+  }
+
+  get hasAllReorderItemsDisabled(): boolean {
+    return (
+      this.hasReorderItems &&
+      this.isTopReorderDisabled &&
+      this.isUpReorderDisabled &&
+      this.isDownReorderDisabled &&
+      this.isBottomReorderDisabled
+    );
+  }
+
+  get hasOnlyDisabledReorderItems(): boolean {
+    return (
+      this.hasReorderItems &&
+      !this.hasMoveToItems &&
+      !this.hasAddToItems &&
+      this.hasAllReorderItemsDisabled
+    );
+  }
+
+  get isTopReorderDisabled(): boolean {
+    const { setPosition } = this;
+
+    return setPosition === 1 || setPosition === 2;
+  }
+
+  get isUpReorderDisabled(): boolean {
+    return this.setPosition === 1;
+  }
+
+  get isDownReorderDisabled(): boolean {
+    return this.setPosition === this.setSize;
+  }
+
+  get isBottomReorderDisabled(): boolean {
+    const { setPosition, setSize } = this;
+
+    return setPosition === setSize || setPosition === setSize - 1;
   }
 
   private interactiveContainer = useInteractive(this);
@@ -104,7 +157,7 @@ export class SortHandle extends LitElement {
   /**
    * Determines where the component will be positioned relative to the container element.
    */
-  @property({ reflect: true }) placement: MenuPlacement = defaultMenuPlacement;
+  @property({ reflect: true }) placement: LogicalPlacement = defaultMenuPlacement;
 
   /** Specifies the size of the component. */
   @property({ reflect: true }) scale: Scale = "m";
@@ -276,6 +329,7 @@ export class SortHandle extends LitElement {
     const {
       disabled,
       flipPlacements,
+      hasOnlyDisabledReorderItems,
       open,
       overlayPositioning,
       placement,
@@ -285,7 +339,7 @@ export class SortHandle extends LitElement {
     } = this;
 
     const text = this.getLabel();
-    const isDisabled = disabled || hasNoItems;
+    const isDisabled = disabled || hasNoItems || hasOnlyDisabledReorderItems;
 
     return (
       <this.interactiveContainer disabled={disabled}>
@@ -353,7 +407,7 @@ export class SortHandle extends LitElement {
   private renderReorderGroup(): JsxNode {
     return this.hasReorderItems ? (
       <calcite-dropdown-group
-        groupTitle={this.messages.reorder}
+        groupTitle={this.reorderGroupTitle}
         id={IDS.reorder}
         key="reorder"
         scale={this.scale}
@@ -368,9 +422,9 @@ export class SortHandle extends LitElement {
   }
 
   private renderAddToGroup(): JsxNode {
-    const { messages, addToItems, scale } = this;
+    const { messages, addToItems, scale, hasAddToItems } = this;
 
-    return addToItems.length ? (
+    return hasAddToItems ? (
       <calcite-dropdown-group
         groupTitle={messages.addTo}
         id={IDS.add}
@@ -384,9 +438,9 @@ export class SortHandle extends LitElement {
   }
 
   private renderMoveToGroup(): JsxNode {
-    const { messages, moveToItems, scale } = this;
+    const { messages, moveToItems, scale, hasMoveToItems } = this;
 
-    return moveToItems.length ? (
+    return hasMoveToItems ? (
       <calcite-dropdown-group
         groupTitle={messages.moveTo}
         id={IDS.move}
@@ -399,10 +453,11 @@ export class SortHandle extends LitElement {
     ) : null;
   }
 
-  private renderDropdownItem(positionIndex: number, label: string): JsxNode {
+  private renderDropdownItem(positionIndex: number, label: string, disabled = false): JsxNode {
     return (
       <calcite-dropdown-item
         data-value={REORDER_VALUES[positionIndex]}
+        disabled={disabled}
         key={REORDER_VALUES[positionIndex]}
         label={label}
         oncalciteDropdownItemSelect={this.handleReorder}
@@ -412,30 +467,20 @@ export class SortHandle extends LitElement {
     );
   }
 
-  private renderTop(): JsxNode | null {
-    const { setPosition } = this;
-
-    return setPosition !== 1 && setPosition !== 2
-      ? this.renderDropdownItem(0, this.messages.moveToTop)
-      : null;
+  private renderTop(): JsxNode {
+    return this.renderDropdownItem(0, this.messages.moveToTop, this.isTopReorderDisabled);
   }
 
-  private renderUp(): JsxNode | null {
-    return this.setPosition !== 1 ? this.renderDropdownItem(1, this.messages.moveUp) : null;
+  private renderUp(): JsxNode {
+    return this.renderDropdownItem(1, this.messages.moveUp, this.isUpReorderDisabled);
   }
 
-  private renderDown(): JsxNode | null {
-    return this.setPosition !== this.setSize
-      ? this.renderDropdownItem(2, this.messages.moveDown)
-      : null;
+  private renderDown(): JsxNode {
+    return this.renderDropdownItem(2, this.messages.moveDown, this.isDownReorderDisabled);
   }
 
-  private renderBottom(): JsxNode | null {
-    const { setPosition, setSize } = this;
-
-    return setPosition !== setSize && setPosition !== setSize - 1
-      ? this.renderDropdownItem(3, this.messages.moveToBottom)
-      : null;
+  private renderBottom(): JsxNode {
+    return this.renderDropdownItem(3, this.messages.moveToBottom, this.isBottomReorderDisabled);
   }
 
   //#endregion